+++ src/Type/ChatMemberMember.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Vjik\TelegramBot\Api\Type;
 
+use DateTimeImmutable;
+
 /**
  * @see https://core.telegram.org/bots/api#chatmembermember
  */
@@ -11,6 +13,7 @@ final readonly class ChatMemberMember implements ChatMember
 {
     public function __construct(
         public User $user,
+        public DateTimeImmutable|null $untilDate = null,
     ) {}
 
     public function getStatus(): string