+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - New #117: Add `businessConnectionId` parameter to `SendPaidMedia` method.
 - New #117: Add `paidMedia` field to `TransactionPartnerUser` type.
+- New #117: Add `subscriptionPeriod` and `subscriptionPrice` fields to `ChatInviteLink` type.
 
 ## 0.4.1 August 1, 2024